fix compile error using ESP-IDF 5.5 and set battery warning level lower to suit computer USB power supply.

+++ esp32/main/mini_pupper_servos.h
@@ -290,7 +290,7 @@ protected:
      * R/W access to register through serial bus using smart protocol
      * 
      */
-    int write_frame(u8 ID, u8 instruction, u8 const * parameters, size_t parameter_length);
+    int write_frame(u8 ID, u8 instruction, u8 const * parameters, u8 parameter_length);
     int reply_frame(u8 & ID, u8 & state, u8 * parameters, size_t parameter_length, TickType_t timeout = 2);
 
     int write_register_byte(u8 id, u8 reg, u8 value);
